Create CardAPIRequestBuilder.
