+++ Card/src/main/java/com/paypal/android/card/network/CardAPIRequestBuilder.kt
@@ -1,0 +1,32 @@
+package com.paypal.android.card.network
+
+import com.paypal.android.card.Card
+import com.paypal.android.core.APIRequest
+import com.paypal.android.core.HttpMethod
+
+class CardAPIRequestBuilder {
+
+    fun buildConfirmPaymentSourceRequest(orderID: String, card: Card): APIRequest {
+        val expirationDate = card.run {
+            val parsedSecurityCode = expirationDate.split("/")
+            val month = parsedSecurityCode[0]
+            val year = parsedSecurityCode[1]
+            "20$year-$month"
+        }
+
+        val cardNumber = card.number.replace("\\s".toRegex(), "")
+        val path = "v2/checkout/orders/$orderID/confirm-payment-source"
+        val body = """
+            {
+                "payment_source": {
+                    "card": {
+                        "number": "$cardNumber",
+                        "expiry": "$expirationDate",
+                        "security_code": "${card.securityCode}"
+                    }
+                }
+            }
+        """.trimIndent()
+        return APIRequest(path, HttpMethod.POST, body)
+    }
+}
